requirements.txt
flask
psycopg2-binary
opencv-python
ultralytics
pytesseract
